Refactor dpor into a header-only interface library
Remove the unused Session API and convert the public dpor target to an INTERFACE library. Add an internal header-check target so project warnings and static analysis still run against the public headers.

Clean up lint findings, normalize repo-wide code formatting, and refresh the affected tests/examples so lint, ASAN, and TSAN validation all pass on the updated tree.

+++ include/dpor/model/exploration_graph.hpp
@@ -497,7 +497,8 @@ class ExplorationGraphT {
     }
 
     std::vector<EventId> kept_ids;
-    kept_ids.reserve(std::count(keep_mask.begin(), keep_mask.end(), std::uint8_t{1}));
+    kept_ids.reserve(
+        static_cast<std::size_t>(std::count(keep_mask.begin(), keep_mask.end(), std::uint8_t{1})));
     std::vector<EventId> id_map(event_count(), kNoSource);
 
     EventId new_id = 0;
